web/requirements.txt
click==6.6
Flask==0.11.1
Flask-SQLAlchemy==2.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.6.2
PyYAML==3.11
SQLAlchemy==1.0.14
Werkzeug==0.11.10